pyproject.toml
[tool.poetry]
name = "team-challenge"
version = "0.1.0"
description = "Dependencies for backend"
authors = ["0leksandrr <tsybulskyy.sasha@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django-import-export = "^3.3.7"
djoser = "^2.2.2"
celery = {extras = ["redis"], version = "^5.3.6"}
flower = "^2.0.1"
sqlparse = "^0.5.0"
gunicorn = "^22.0.0"
requests = "^2.32.2"
social-auth-app-django = "^5.4.1"
idna = "^3.7"
pillow = "^10.3.0"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "2.9.9"
python-dotenv = "1.0.1"
waitress = "3.0.0"
django-filter = "24.1"
drf-yasg = "1.21.7"
django = "4.2.11"
django-cors-headers = "4.3.1"
djangorestframework = "3.14.0"
pytz = "2023.3.post1"
tzdata = "2023.4"
django-extensions = "3.2.3"
django-storages = "1.14.2"

[tool.poetry.group.dev.dependencies]
asgiref = "3.7.2"
cachetools = "5.3.2"
certifi = "2024.2.2"
charset-normalizer = "3.3.2"
google-api-core = "2.17.0"
google-auth = "2.27.0"
google-cloud-core = "2.4.1"
google-cloud-storage = "2.14.0"
google-crc32c = "1.5.0"
google-resumable-media = "2.7.0"
googleapis-common-protos = "1.62.0"
inflection = "0.5.1"
packaging = "23.2"
protobuf = "4.25.2"
pyasn1 = "0.5.1"
pyasn1-modules = "0.3.0"
pyyaml = "6.0.1"
rsa = "4.9"
setuptools = "69.0.3"
uritemplate = "4.1.1"
urllib3 = "2.2.1"
coverage = "^7.5.2"
pip = "^24.0"
install = "^1.3.5"
factory-boy = "^3.3.0"
isort = "^5.13.2"
pre-commit = "^3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 120
ignore = ["E203", "E266", "E501", "W503"]
select = ["B", "C", "E", "F", "W", "T4", "B9"]


[tool.isort]
multi_line_output=3
include_trailing_comma=true
line_length=120
lines_after_imports=2
lines_between_sections=1
lines_between_types=0
force_alphabetical_sort_within_sections=true
ensure_newline_before_comments=true
combine_as_imports=true
combine_star=true
skip_gitignore=true
balanced_wrapping=true
force_grid_wrap=2
src_paths=["core", ]
skip_glob=[
    ".git",
    ".idea",
    ".vscode",
    "venv",
    ".venv",
    "*migrations*",
]
known_django=["django","rest_framework"]
known_first_party=["core"]
sections=[
    "FUTURE",
    "STDLIB",
    "DJANGO",
    "THIRDPARTY",
    "FIRSTPARTY",
    "LOCALFOLDER",
]